backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
chromadb
pypdf
python-dotenv
tiktoken
